refactor: separate the Comment module
