+++ src/articles/articles.module.ts
@@ -5,6 +5,7 @@ import { IpfsModule } from "src/ipfs/ipfs.module";
 
 import { ArticlesController } from "./articles.controller";
 import { ArticlesService } from "./articles.service";
+import { CommentsModule } from "./comments/comments.module";
 import { Article } from "./entities/article.entity";
 import { Comment } from "./entities/comment.entity";
 
@@ -13,6 +14,7 @@ import { Comment } from "./entities/comment.entity";
     HttpModule,
     IpfsModule,
     TypeOrmModule.forFeature([Article, Comment]),
+    CommentsModule,
   ],
   controllers: [ArticlesController],
   providers: [ArticlesService],
