+++ src/articles/comments/comments.controller.spec.ts
@@ -1,0 +1,21 @@
+import { Test, TestingModule } from "@nestjs/testing";
+
+import { CommentsController } from "./comments.controller";
+import { CommentsService } from "./comments.service";
+
+describe("CommentsController", () => {
+  let controller: CommentsController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [CommentsController],
+      providers: [CommentsService],
+    }).compile();
+
+    controller = module.get<CommentsController>(CommentsController);
+  });
+
+  it("should be defined", () => {
+    expect(controller).toBeDefined();
+  });
+});
