+++ src/articles/comments/comments.module.ts
@@ -1,0 +1,13 @@
+import { Module } from "@nestjs/common";
+import { TypeOrmModule } from "@nestjs/typeorm";
+
+import { Comment } from "../entities/comment.entity";
+import { CommentsController } from "./comments.controller";
+import { CommentsService } from "./comments.service";
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Comment])],
+  controllers: [CommentsController],
+  providers: [CommentsService],
+})
+export class CommentsModule {}
