+++ src/articles/comments/comments.service.spec.ts
@@ -1,0 +1,19 @@
+import { Test, TestingModule } from "@nestjs/testing";
+
+import { CommentsService } from "./comments.service";
+
+describe("CommentsService", () => {
+  let service: CommentsService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [CommentsService],
+    }).compile();
+
+    service = module.get<CommentsService>(CommentsService);
+  });
+
+  it("should be defined", () => {
+    expect(service).toBeDefined();
+  });
+});
